Refactors all the Statement classes and subclasses in a new statement.py module
